move import to ranking object and fix listing controller and view
