+++ db/schema.rb
@@ -10,29 +10,15 @@
 #
 # It's strongly recommended that you check this file into your version control system.
 
-ActiveRecord::Schema.define(version: 2020_04_26_091347) do
+ActiveRecord::Schema.define(version: 2021_09_24_184238) do
 
-  create_table "clubs", force: :cascade do |t|
-    t.integer "dtb_id"
-    t.string "club"
-    t.string "federation"
-    t.index ["dtb_id"], name: "index_clubs_on_dtb_id"
-  end
-
-  create_table "players", force: :cascade do |t|
+  create_table "rankings", force: :cascade do |t|
     t.integer "dtb_id"
     t.string "firstname", limit: 50
     t.string "lastname", limit: 50
     t.string "federation", limit: 50
     t.string "club"
     t.string "nationality", limit: 3
-    t.datetime "created_at", precision: 6, null: false
-    t.datetime "updated_at", precision: 6, null: false
-    t.index ["dtb_id"], name: "index_players_on_dtb_id"
-  end
-
-  create_table "rankings", force: :cascade do |t|
-    t.integer "dtb_id"
     t.date "date"
     t.string "age_group"
     t.integer "ranking_position"
@@ -42,10 +28,12 @@ ActiveRecord::Schema.define(version: 2020_04_26_091347) do
     t.boolean "year_end_ranking", default: false
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
-    t.string "club"
-    t.string "federation"
+    t.index ["club"], name: "index_rankings_on_club"
+    t.index ["dtb_id", "date", "federation"], name: "index_rankings_on_dtb_id_and_date_and_federation"
     t.index ["dtb_id", "date"], name: "index_rankings_on_dtb_id_and_date"
     t.index ["dtb_id"], name: "index_rankings_on_dtb_id"
+    t.index ["federation"], name: "index_rankings_on_federation"
+    t.index ["lastname"], name: "index_rankings_on_lastname"
   end
 
   create_table "users", force: :cascade do |t|
